pyproject.toml
[project]
name = "flow-guided-krylov"
version = "0.1.0"
description = "Flow-Guided Krylov Quantum Diagonalization for Ground State Energy Computation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Flow-Guided-Krylov Contributors"}
]
keywords = [
    "quantum computing",
    "neural quantum states",
    "normalizing flows",
    "krylov methods",
    "quantum chemistry",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Chemistry",
]
dependencies = [
    "torch>=2.0.0",
    "normflows>=1.7.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "pyscf>=2.3.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
cuda = [
    "cupy-cuda12x>=12.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
]

[project.urls]
Homepage = "https://github.com/flow-guided-krylov/flow-guided-krylov"
Documentation = "https://flow-guided-krylov.readthedocs.io"
Repository = "https://github.com/flow-guided-krylov/flow-guided-krylov"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
